requirements.txt
django4.2
python3.11
dajngo-extensions
celery
django-debug-toolbar
djangorestframework